requirements.txt
streamlit
python-dotenv
requests
pinecone-client
numpy
docusign-esign